+++ tvos/g33kVaultTV/g33kVaultTV/g33kVaultTVApp.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+
+@main
+struct g33kVaultTVApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}